# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateCartItemDTO:
    properties:
      product_id:
        type: integer
    required:
    - product_id
    type: object
  dto.CreateCategoryDTO:
    properties:
      category_name:
        maxLength: 50
        minLength: 5
        type: string
      category_slug:
        maxLength: 60
        minLength: 5
        type: string
    required:
    - category_name
    - category_slug
    type: object
...
  dto.GetCartItemDTO:
    properties:
      cart_id:
        type: integer
      cart_item_id:
        type: integer
      product:
        $ref: '#/definitions/dto.GetProductDTO'
      quantity:
        type: integer
    type: object
  dto.GetCategoryDTO:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      category_slug:
        type: string
    type: object
...
  dto.GetProductDTO:
    properties:
      category:
        $ref: '#/definitions/dto.GetCategoryDTO'
      product_description:
        type: string
      product_id:
        type: integer
      product_image:
        type: string
      product_name:
        type: string
      product_price:
        type: number
      seller_id:
        type: integer
    type: object
  dto.LoginDTO:
    properties:
      password:
        maxLength: 60
        minLength: 8
        type: string
      username:
        maxLength: 40
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  handler.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      roles:
        example:
        - user
        - seller
        items:
          type: string
        type: array
    type: object
  handler.CartItemIDResponse:
    properties:
      cart_item_id:
        example: 33
        type: integer
    type: object
  handler.CategoryIDResponse:
    properties:
      category_id:
        example: 7
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      message:
        example: validation failed
        type: string
    type: object
  handler.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  handler.QuantityResponse:
    properties:
      quantity:
        example: 3
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 'REST API for marketplace: products, categories, cart, orders, auth
    and roles.'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Marketplace API
  version: "1.0"
paths:
  /cart:
    get:
      description: Get all items in the authenticated user's cart.
      produces:
      - application/json
      responses:
        "200":
          description: List of cart items
          schema:
            items:
              $ref: '#/definitions/dto.GetCartItemDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List cart items
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Add a product to the authenticated user's cart.
      parameters:
      - description: Product to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCartItemDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Created cart item ID
          schema:
            $ref: '#/definitions/handler.CartItemIDResponse'
        "400":
          description: Parse or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/items/{cart_item_id}:
    delete:
      description: Delete a cart item by ID.
      parameters:
      - description: Cart item ID
        in: path
        name: cart_item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid cart_item_id
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    patch:
      description: Change quantity of a cart item by delta (can be negative).
      parameters:
      - description: Cart item ID
        in: path
        name: cart_item_id
        required: true
        type: integer
      - description: Quantity change (positive or negative)
        in: query
        name: delta
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Updated quantity
          schema:
            $ref: '#/definitions/handler.QuantityResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Cart item not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
  /categories:
    get:
      description: Get all product categories.
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            items:
              $ref: '#/definitions/dto.GetCategoryDTO'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category. Requires admin role.
      parameters:
      - description: Category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Created category ID
          schema:
            $ref: '#/definitions/handler.CategoryIDResponse'
        "400":
          description: Parse or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden (not admin)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Category name or slug already exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Categories
...
  /login:
    post:
      consumes:
      - application/json
      description: 'Authenticate user by username and password. Returns access token
        and sets refresh token cookie. Rate limit: 5 req/min.'
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Access token and user roles
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "400":
          description: Parse or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /logout:
    post:
      description: Invalidate refresh and access tokens. Requires valid access token.
        Clears refresh token cookie.
      produces:
      - application/json
      responses:
        "200":
          description: Logout confirmation
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "401":
          description: Unauthorized or missing refresh token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
...
  /refresh:
    post:
      description: 'Refresh access token using refresh token from cookie. Rate limit:
        3 req/min.'
      produces:
      - application/json
      responses:
        "200":
          description: New access token and user roles
          schema:
            $ref: '#/definitions/handler.AuthResponse'
        "401":
          description: Missing or invalid refresh token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Refresh tokens
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"